Prevent removing resources marked as used

+++ src/aircraft/aircraft.service.ts
@@ -8,6 +8,7 @@ import { UpdateAircraftDto } from './dto/update-aircraft.dto';
 import { PrismaService } from '../prisma/prisma.service';
 import { v4 } from 'uuid';
 import { Aircraft } from '@prisma/client';
+import { AircraftInUseError } from './dto/errors.dto';
 
 @Injectable()
 export class AircraftService {
@@ -59,6 +60,14 @@ export class AircraftService {
 
     if (!aircraft) {
       throw new NotFoundException('Aircraft with given id does not exist.');
+    }
+
+    const connectedFlights = await this.prisma.flight.count({
+      where: { aircraftId: id },
+    });
+
+    if (connectedFlights > 0) {
+      throw new BadRequestException(AircraftInUseError);
     }
 
     await this.prisma.aircraft.delete({ where: { id } });
